Move views into visualization folder
