+++ js/visualization/views/mcrecoassociation.js
