+++ js/visualization/views/templates/association-view.js
