+++ js/visualization/views/templates/list.js
