+++ js/visualization/views/templates/mcparticletree.js
