+++ js/visualization/views/templates/recoclustertrack.js
