+++ js/visualization/views/templates/tree.js
