+++ js/visualization/views/vertexlist.js
@@ -1,5 +1,5 @@
 import { listView } from "./templates/list.js";
-import { preFilterList } from "../filters/pre-filter.js";
+import { preFilterList } from "../../filters/pre-filter.js";
 
 export function vertexList(viewCurrentObjects) {
   const vertexCollection =
